requirements.txt
blinker==1.6.2
boto3==1.28.12
botocore==1.31.12
chardet==5.1.0
click==8.1.6
contourpy==1.1.0
cycler==0.11.0
distro==1.8.0
et-xmlfile==1.1.0
Flask==2.3.2
fonttools==4.41.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.25.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pdfminer.six==20191110
Pillow==10.0.0
pycryptodome==3.18.0
pyparsing==3.0.9
pypdf2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sortedcontainers==2.4.0
tabula-py==2.7.0
typing-extensions==4.7.1
tzdata==2023.3
UliPlot==0.2.2
urllib3==1.26.16
Werkzeug==2.3.6
zipp==3.16.2
